split up operations.py for better organization
