+++ server/app/llm/json_parser.py
@@ -1,0 +1,40 @@
+import json
+import re
+from typing import Dict
+
+
+class JSONParser:
+    """Handles JSON parsing and validation from LLM responses"""
+
+    @staticmethod
+    def validate_and_extract_json(json_data: str) -> dict:
+        """Extract and validate JSON data from various formats"""
+        if not json_data or not isinstance(json_data, str):
+            raise ValueError("Invalid input: empty or non-string data")
+
+        json_data = json_data.strip()
+
+        # Case 1: Try parsing directly first
+        try:
+            return json.loads(json_data)
+        except json.JSONDecodeError:
+            pass
+
+        # Case 2: Check for code block format
+        if "```" in json_data:
+            code_blocks = re.findall(r"```(?:json)?\s*([\s\S]*?)```", json_data)
+
+            for block in code_blocks:
+                block = block.strip()
+                block = re.sub(r"}\s+\w+\s+}", "}}", block)
+                block = re.sub(r"}\s+\w+\s+,", "},", block)
+
+                try:
+                    return json.loads(block)
+                except json.JSONDecodeError:
+                    continue
+
+        raise ValueError(
+            "Could not extract valid JSON from the provided string. "
+            "Please ensure the response contains proper JSON format."
+        )
