+++ server/app/llm/operations.py
@@ -1,624 +1,24 @@
-import json
-import logging
-import os
-import re
-import uuid
-from typing import AsyncGenerator, Optional, Sequence, Union
-
-import httpx
-from app.database.crud.message_crud import message_crud
-from app.database.crud.paper_crud import paper_crud
-from app.database.database import get_db
-from app.database.models import Message, Paper
-from app.helpers.s3 import s3_service
-from app.llm.prompts import (
-    ANSWER_PAPER_QUESTION_SYSTEM_PROMPT,
-    ANSWER_PAPER_QUESTION_USER_MESSAGE,
-    CONCISE_MODE_INSTRUCTIONS,
-    DETAILED_MODE_INSTRUCTIONS,
-    EXTRACT_PAPER_METADATA,
-    EXTRACT_RELEVANT_CONTEXT_FROM_PAPER,
-    EXTRACT_RELEVANT_FINDINGS_FROM_PAPER_SUMMARIES,
-    FINALIZE_HYPOTHESIS_RESEARCH,
-    HYPOTHESIS_STEPS,
-    IMPROVE_QUESTION_TO_HYPOTHESIS,
-    NORMAL_MODE_INSTRUCTIONS,
-    PICK_PAPERS_TO_READ,
-)
-from app.llm.schemas import (
-    HypothesisFanOut,
-    HypothesisResearchResponse,
-    HypothesisStep,
-    MinimalPaperData,
-    PaperMetadataExtraction,
-    PapersForReference,
-    WhatToScrape,
-)
-from app.llm.utils import retry_llm_operation
-from app.schemas.message import ResponseStyle
-from app.schemas.user import CurrentUser
-from fastapi import Depends
-from google import genai  # type: ignore
-from google.genai.types import Content, Part  # type: ignore
-from sqlalchemy.orm import Session
-
-GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")
-
-gemini_client = genai.Client(api_key=GEMINI_API_KEY)
-
-logger = logging.getLogger(__name__)
+from app.llm.citation_handler import CitationHandler
+from app.llm.hypothesis_operations import HypothesisOperations
+from app.llm.json_parser import JSONParser
+from app.llm.paper_operations import PaperOperations
 
 
-class Operations:
+# For backward compatibility, create a unified Operations class
+class Operations(HypothesisOperations, PaperOperations):
     """
-    Class to handle operations related to LLM
+    Unified operations class that combines all LLM operations
+    Inherits from both hypothesis and paper operations for backward compatibility
     """
 
-    def __init__(self):
-        self.client = genai.Client(api_key=GEMINI_API_KEY)
-        self.default_model = "gemini-2.5-pro-preview-03-25"
-        self.fast_model = "gemini-2.5-flash-preview-04-17"
-
-    def validate_and_extract_json(self, json_data: str) -> dict:
-        """
-        Extract and validate JSON data from various formats
-
-        Args:
-            json_data (str): String which may contain JSON in different formats
-
-        Returns:
-            dict: Parsed JSON data
-
-        Raises:
-            ValueError: If valid JSON cannot be extracted
-        """
-        if not json_data or not isinstance(json_data, str):
-            raise ValueError("Invalid input: empty or non-string data")
-
-        # Remove any leading/trailing whitespace
-        json_data = json_data.strip()
-
-        # Case 1: Try parsing directly first
-        try:
-            return json.loads(json_data)
-        except json.JSONDecodeError:
-            pass
-
-        # Case 2: Check for code block format (```json ... ```)
-        if "```" in json_data:
-            # Find content between triple backticks
-            import re
-
-            code_blocks = re.findall(r"```(?:json)?\s*([\s\S]*?)```", json_data)
-
-            # Try each code block
-            for block in code_blocks:
-                block = block.strip()
-
-                # Clean up common LLM artifacts that break JSON
-                # Remove standalone words that appear after closing braces/brackets
-                block = re.sub(r"}\s+\w+\s+}", "}}", block)  # Fix case like "} word }"
-                block = re.sub(r"}\s+\w+\s+,", "},", block)  # Fix case like "} word ,"
-
-                try:
-                    return json.loads(block)
-                except json.JSONDecodeError:
-                    continue
-
-        # If we got here, we couldn't find valid JSON
-        raise ValueError(
-            "Could not extract valid JSON from the provided string. "
-            "Please ensure the response contains proper JSON format."
-        )
-
-    def convert_references_to_dict(self, references: Sequence[str]) -> dict:
-        """
-        Convert user references to structured citations
-        """
-        citations = []
-        for idx, ref in enumerate(references):
-            citations.append(
-                {
-                    "key": idx + 1,
-                    "reference": ref,
-                }
-            )
-        return {
-            "citations": citations,
-        }
-
-    def convert_references_to_citations(
-        self, references: Optional[Sequence[str]]
-    ) -> str:
-        """
-        Convert user references to structured citations
-        """
-        if not references:
-            return ""
-
-        return self.format_citations(
-            self.convert_references_to_dict(references)["citations"]
-        )
-
-    def format_citations(
-        self,
-        citations: list[dict],
-    ) -> str:
-        """
-        Format citations into a structured string
-        """
-        citation_format = "---EVIDENCE---\n"
-
-        citation_format += "\n".join(
-            [
-                f"@cite[{citation['key']}]\n{citation['reference']}"
-                for citation in citations
-            ]
-        )
-
-        citation_format += "\n---END-EVIDENCE---"
-        return citation_format
-
-    def convert_chat_history_to_api_format(
-        self,
-        messages: list[Message],
-    ) -> list[Content]:
-        """
-        Convert chat history to Chat API format
-        """
-        api_format = []
-        for message in messages:
-            references = self.format_citations(message.references["citations"]) if message.references else None  # type: ignore
-
-            f_message = (
-                f"{message.content}\n\n{references}" if references else message.content
-            )
-
-            api_format.append(
-                Content(
-                    role="user" if message.role == "user" else "model",
-                    parts=[{"text": f_message}],  # type: ignore
-                )
-            )
-        return api_format
-
-    async def explain_text(self, contents: str, model: str = "gemini-2.0-flash"):
-        """
-        Explain the provided text using the specified model
-        """
-        async for chunk in self.client.models.generate_content_stream(  # type: ignore
-            model=model, contents=contents
-        ):
-            # Process the chunk of generated content
-            yield chunk.text
-
-    @retry_llm_operation(max_retries=3, delay=1.0)
-    def augment_hypothesis(
-        self,
-        question: str,
-    ) -> str:
-        """
-        Augment the provided hypothesis using the specified model
-        """
-        formatted_prompt = IMPROVE_QUESTION_TO_HYPOTHESIS.format(
-            question=question,
-        )
-
-        # Generate content using the LLM
-        augmented_hypothesis = self.client.models.generate_content(
-            model=self.fast_model,
-            contents=formatted_prompt,
-        )
-
-        # Check if the response is valid JSON
-        if augmented_hypothesis and augmented_hypothesis.text:
-            return augmented_hypothesis.text.strip()
-        raise ValueError("Failed to generate augmented hypothesis from LLM.")
-
-    @retry_llm_operation(max_retries=3, delay=1.0)
-    def generate_steps(
-        self,
-        hypothesis: str,
-    ) -> HypothesisFanOut:
-        """
-        Generate steps to test the hypothesis using the specified model
-        """
-        # Construct the expected schema for the response
-        schema = HypothesisFanOut.model_json_schema()
-
-        # Format the prompt with the hypothesis and schema
-        formatted_prompt = HYPOTHESIS_STEPS.format(
-            hypothesis=hypothesis,
-            schema=schema,
-        )
-
-        # Generate content using the LLM
-        response = self.client.models.generate_content(
-            model=self.fast_model,
-            contents=formatted_prompt,
-        )
-
-        # Check if the response is valid JSON
-        try:
-            if response and response.text:
-                response_json = self.validate_and_extract_json(response.text)
-                # Validate the response against the schema
-                return HypothesisFanOut.model_validate(response_json)
-            else:
-                raise ValueError("Empty response from LLM.")
-        except ValueError as e:
-            raise ValueError(f"Invalid JSON response from LLM: {str(e)}")
-
-    @retry_llm_operation(max_retries=3, delay=1.0)
-    def select_papers_to_read(
-        self,
-        hypothesis: str,
-        question: str,
-        motivation: str,
-        minimal_papers: Sequence[MinimalPaperData],
-    ) -> WhatToScrape:
-        schema = WhatToScrape.model_json_schema()
-        formatted_prompt = PICK_PAPERS_TO_READ.format(
-            hypothesis=hypothesis,
-            question=question,
-            motivation=motivation,
-            minimal_papers=minimal_papers,
-            schema=schema,
-        )
-
-        response = self.client.models.generate_content(
-            model=self.fast_model,
-            contents=formatted_prompt,
-        )
-        # Check if the response is valid JSON
-        try:
-            if response and response.text:
-                response_json = self.validate_and_extract_json(response.text)
-                # Validate the response against the schema
-                return WhatToScrape.model_validate(response_json)
-            else:
-                raise ValueError("Empty response from LLM while selecting papers.")
-        except ValueError as e:
-            raise ValueError(f"Invalid JSON response from LLM: {str(e)}")
-
-    @retry_llm_operation(max_retries=3, delay=1.0)
-    def summarize_paper(
-        self,
-        hypothesis: str,
-        question: str,
-        motivation: str,
-        paper: str,
-    ) -> str | None:
-        """
-        Summarize the paper using the specified model
-        """
-        formatted_prompt = EXTRACT_RELEVANT_CONTEXT_FROM_PAPER.format(
-            hypothesis=hypothesis,
-            question=question,
-            motivation=motivation,
-            paper=paper,
-        )
-
-        response = self.client.models.generate_content(
-            model=self.fast_model,
-            contents=formatted_prompt,
-        )
-
-        if response and response.text:
-            if response.text.strip() == "Skip":
-                logger.warning(
-                    f"Paper skipped for hypothesis: {hypothesis}, question: {question}"
-                )
-                return None
-            return response.text.strip()
-        else:
-            raise ValueError(
-                f"Empty response from LLM while summarizing paper: {paper[:50]}..."
-            )
-
-    @retry_llm_operation(max_retries=3, delay=1.0)
-    def collate_findings(
-        self,
-        hypothesis: str,
-        step: HypothesisStep,
-        papers: Sequence[PapersForReference],
-    ) -> str:
-        """
-        Collate findings from the papers for the given hypothesis step
-        """
-        formatted_prompt = EXTRACT_RELEVANT_FINDINGS_FROM_PAPER_SUMMARIES.format(
-            hypothesis=hypothesis,
-            question=step.question,
-            motivation=step.motivation,
-            paper_summaries=[paper.model_dump_json() for paper in papers],
-        )
-
-        response = self.client.models.generate_content(
-            model=self.fast_model,
-            contents=formatted_prompt,
-        )
-
-        if response and response.text:
-            return response.text.strip()
-        else:
-            raise ValueError("Empty response from LLM while collating findings.")
-
-    @retry_llm_operation(max_retries=3, delay=1.0)
-    def finalize_hypothesis_research(
-        self,
-        hypothesis: str,
-        steps: Sequence[dict],
-    ) -> HypothesisResearchResponse:
-        """
-        Finalize the hypothesis research by collating findings from all steps.
-        """
-        formatted_prompt = FINALIZE_HYPOTHESIS_RESEARCH.format(
-            hypothesis=hypothesis,
-            steps_results=steps,
-            schema=HypothesisResearchResponse.model_json_schema(),
-        )
-
-        response = self.client.models.generate_content(
-            model=self.default_model,
-            contents=formatted_prompt,
-        )
-
-        try:
-            if response and response.text:
-                response_json = self.validate_and_extract_json(response.text)
-                # Validate the response against the schema
-                return HypothesisResearchResponse.model_validate(response_json)
-            else:
-                raise ValueError(
-                    "Empty response from LLM while finalizing hypothesis research."
-                )
-        except ValueError as e:
-            raise ValueError(f"Invalid JSON response from LLM: {str(e)}")
-
-    @retry_llm_operation(max_retries=3, delay=1.0)
-    def extract_paper_metadata(
-        self,
-        paper_id: str,
-        user: CurrentUser,
-        file_path: Optional[str] = None,
-        db: Session = Depends(get_db),
-    ) -> PaperMetadataExtraction:
-        """
-        Extract metadata from the paper using the specified model
-        """
-        paper = paper_crud.get(db, id=paper_id, user=user)
-
-        if not paper:
-            raise ValueError(f"Paper with ID {paper_id} not found.")
-
-        # Load and extract raw data from the PDF
-        raw_file = paper_crud.read_raw_document_content(
-            db, paper_id=paper_id, current_user=user, file_path=file_path
-        )
-
-        if not raw_file:
-            raise ValueError(
-                f"Raw file content for paper ID {paper_id} could not be retrieved."
-            )
-
-        formatted_prompt = EXTRACT_PAPER_METADATA.format(
-            paper=raw_file, schema=PaperMetadataExtraction.model_json_schema()
-        )
-
-        # Extract metadata using the LLM
-        response = self.client.models.generate_content(
-            model=self.fast_model,
-            contents=formatted_prompt,
-        )
-
-        # Check if the response is valid JSON
-        try:
-            if response and response.text:
-                response_json = self.validate_and_extract_json(response.text)
-        except ValueError as e:
-            raise ValueError(f"Invalid JSON response from LLM: {str(e)}")
-
-        # Parse the response and return the metadata
-        metadata = PaperMetadataExtraction.model_validate(response_json)
-        return metadata
-
-    async def chat_with_paper(
-        self,
-        paper_id: str,
-        conversation_id: str,
-        question: str,
-        current_user: CurrentUser,
-        user_references: Optional[Sequence[str]] = None,
-        response_style: Optional[str] = "normal",
-        db: Session = Depends(get_db),
-    ) -> AsyncGenerator[Union[str, dict], None]:
-        """
-        Chat with the paper using the specified model
-        """
-
-        def parse_evidence_block(evidence_text: str) -> list[dict]:
-            """
-            Parse evidence block into structured citations
-            Handles multi-line citations between @cite markers
-
-            Incoming format of evidence_text:
-            @cite[1]
-            "First piece of evidence"
-            @cite[2]
-            "Second piece of evidence"
-            """
-            citations = []
-            lines = evidence_text.strip().split("\n")
-            current_citation: dict[str, Union[int, str]] | None = None
-            current_text_lines: list[str] = []
-
-            for line in lines:
-                line = line.strip()
-                if line.startswith("@cite["):
-                    # If we have a previous citation pending, save it
-                    if current_citation is not None:
-                        current_citation["reference"] = " ".join(
-                            current_text_lines
-                        ).strip()
-                        citations.append(current_citation)
-
-                    # Start new citation
-                    match = re.search(r"@cite\[(\d+)\]", line)
-                    if match:
-                        number = int(match.group(1))
-                        current_citation = {"key": number, "reference": ""}
-                        current_text_lines = []
-                elif current_citation is not None and line:
-                    # Accumulate lines for the current citation
-                    current_text_lines.append(line)
-
-            # Don't forget to save the last citation
-            if current_citation is not None and current_text_lines:
-                current_citation["reference"] = " ".join(current_text_lines).strip()
-                citations.append(current_citation)
-
-            return citations
-
-        user_citations = (
-            self.convert_references_to_citations(user_references)
-            if user_references
-            else None
-        )
-
-        paper: Paper = paper_crud.get(db, id=paper_id)
-
-        if not paper:
-            raise ValueError(f"Paper with ID {paper_id} not found.")
-
-        casted_conversation_id = uuid.UUID(conversation_id)
-
-        conversation_history = message_crud.get_conversation_messages(
-            db, conversation_id=casted_conversation_id, current_user=current_user
-        )
-
-        chat_history = self.convert_chat_history_to_api_format(conversation_history)
-
-        additional_instructions = ""
-
-        if response_style == ResponseStyle.DETAILED:
-            additional_instructions = DETAILED_MODE_INSTRUCTIONS
-        elif response_style == ResponseStyle.CONCISE:
-            additional_instructions = CONCISE_MODE_INSTRUCTIONS
-        else:
-            additional_instructions = NORMAL_MODE_INSTRUCTIONS
-
-        formatted_system_prompt = ANSWER_PAPER_QUESTION_SYSTEM_PROMPT.format(
-            additional_instructions=additional_instructions,
-        )
-
-        chat_session = self.client.chats.create(
-            model=self.default_model,
-            history=chat_history,
-            config={
-                "system_instruction": formatted_system_prompt,
-            },
-        )
-
-        formatted_prompt = ANSWER_PAPER_QUESTION_USER_MESSAGE.format(
-            question=f"{question}\n\n{user_citations}" if user_citations else question,
-        )
-
-        evidence_buffer: list[str] = []
-        text_buffer: str = ""
-        in_evidence_section = False
-
-        START_DELIMITER = "---EVIDENCE---"
-        END_DELIMITER = "---END-EVIDENCE---"
-
-        signed_url = s3_service.generate_presigned_url(
-            object_key=str(paper.s3_object_key)
-        )
-
-        if not signed_url:
-            raise ValueError(
-                f"Could not generate presigned URL for paper with ID {paper_id}."
-            )
-
-        # Retrieve and encode the PDF byte
-        pdf_bytes = httpx.get(signed_url).content
-
-        message_content = [
-            Part.from_bytes(
-                data=pdf_bytes,
-                mime_type="application/pdf",
-            ),
-            formatted_prompt,
-        ]
-
-        # Extract metadata using the LLM
-        for chunk in chat_session.send_message_stream(
-            message=message_content,
-        ):
-            text = chunk.text
-
-            logger.debug(f"Received chunk: {text}")
-
-            if not text:
-                continue
-
-            text_buffer += text
-
-            # Check for start delimiter
-            if not in_evidence_section and START_DELIMITER in text_buffer:
-                in_evidence_section = True
-                # Split at delimiter and yield any content that came before
-                pre_evidence = text_buffer.split(START_DELIMITER)[0]
-                if pre_evidence:
-                    yield {"type": "content", "content": pre_evidence}
-                # Start the evidence buffer
-                evidence_buffer = [text_buffer.split(START_DELIMITER)[1]]
-                # Clear the text buffer
-                text_buffer = ""
-                continue
-
-            reconstructed_buffer = "".join(evidence_buffer + [text_buffer]).strip()
-
-            if in_evidence_section and END_DELIMITER in reconstructed_buffer:
-                # Split at delimiter
-                evidence_part, remaining = (
-                    text_buffer.split(END_DELIMITER)
-                    if END_DELIMITER in text_buffer
-                    else ("", "")
-                )
-                evidence_buffer.append(evidence_part)
-
-                # Parse the complete evidence block
-                raw_evidence = "".join(evidence_buffer).strip()
-                structured_evidence = parse_evidence_block(raw_evidence)
-
-                # Yield both raw and structured evidence
-                yield {
-                    "type": "references",
-                    "content": {
-                        "citations": structured_evidence,
-                    },
-                }
-
-                # Reset buffers and state
-                in_evidence_section = False
-                evidence_buffer = []
-                text_buffer = remaining
-
-                # Yield any remaining content after evidence section
-                if remaining:
-                    yield {"type": "content", "content": remaining}
-                continue
-
-            # Handle normal streaming
-            if in_evidence_section:
-                evidence_buffer.append(text)
-                text_buffer = ""
-            else:
-                # Keep a reasonable buffer size for detecting delimiters
-                if len(text_buffer) > len(START_DELIMITER) * 2:
-                    to_yield = text_buffer[: -len(START_DELIMITER)]
-                    yield {"type": "content", "content": to_yield}
-                    text_buffer = text_buffer[-len(START_DELIMITER) :]
-
-        if text_buffer:
-            yield {"type": "content", "content": text_buffer}
+    pass
+
+
+# Also expose individual components for more targeted usage
+__all__ = [
+    "Operations",
+    "HypothesisOperations",
+    "PaperOperations",
+    "CitationHandler",
+    "JSONParser",
+]
